+++ src/com/intellij/idea/plugin/hybris/common/utils/HybrisIcons.kt
@@ -249,6 +249,8 @@ object HybrisIcons {
     val CCV2_SERVICE = AllIcons.Nodes.Services
     val CCV2_SERVICE_MODIFIED_BY = AllIcons.General.User
 
+    val CCV2_DATA_BACKUP_CREATED_BY = AllIcons.General.User
+
     val CCV2_DEPLOYMENT_CREATED_BY = AllIcons.General.User
     val CCV2_DEPLOYMENT_UPDATE_MODE_NONE = AllIcons.Diff.GutterCheckBox
     val CCV2_DEPLOYMENT_UPDATE_MODE_UPDATE = ExpUiIcons.General.PluginUpdate